refactor: split attachment options
